+++ src/api-client.ts
@@ -85,6 +85,7 @@ export async function getGitHubVersionFromApi(
 
   // Doesn't strictly have to be the meta endpoint as we're only
   // using the response headers which are available on every request.
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-call
   const response = await apiClient.rest.meta.get();
 
   // This happens on dotcom, although we expect to have already returned in that
@@ -138,9 +139,18 @@ export async function getWorkflowRelativePath(): Promise<string> {
   );
   const workflowUrl = runsResponse.data.workflow_url;
 
+  const requiredWorkflowRegex =
+    /\/repos\/[^/]+\/[^/]+\/actions\/required_workflows\/[^/]+/;
+  if (!workflowUrl || requiredWorkflowRegex.test(workflowUrl as string)) {
+    // For required workflows, the workflowUrl is invalid so we cannot fetch more informations
+    // about the workflow.
+    // However, the path is available in the original response.
+    return runsResponse.data.path as string;
+  }
+
   const workflowResponse = await apiClient.request(`GET ${workflowUrl}`);
 
-  return workflowResponse.data.path;
+  return workflowResponse.data.path as string;
 }
 
 /**