+++ src/codeql.ts
@@ -233,14 +233,14 @@ export interface BetterResolveLanguagesOutput {
 export interface ResolveQueriesOutput {
   byLanguage: {
     [language: string]: {
-      [queryPath: string]: {};
+      [queryPath: string]: object;
     };
   };
   noDeclaredLanguage: {
-    [queryPath: string]: {};
+    [queryPath: string]: object;
   };
   multipleDeclaredLanguages: {
-    [queryPath: string]: {};
+    [queryPath: string]: object;
   };
 }
 
@@ -422,9 +422,9 @@ function resolveFunction<T>(
     const dummyMethod = () => {
       throw new Error(`CodeQL ${methodName} method not correctly defined`);
     };
-    return dummyMethod as any;
+    return dummyMethod as T;
   }
-  return partialCodeql[methodName];
+  return partialCodeql[methodName] as T;
 }
 
 /**
@@ -612,12 +612,19 @@ export async function getCodeQLForCmd(
         extraArgs.push("--no-sublanguage-file-coverage");
       }
 
+      const overwriteFlag = isSupportedToolsFeature(
+        await this.getVersion(),
+        ToolsFeature.ForceOverwrite,
+      )
+        ? "--force-overwrite"
+        : "--overwrite";
+
       await runTool(
         cmd,
         [
           "database",
           "init",
-          "--overwrite",
+          overwriteFlag,
           "--db-cluster",
           config.dbLocation,
           `--source-root=${sourceRoot}`,
@@ -681,6 +688,8 @@ export async function getCodeQLForCmd(
           "database",
           "trace-command",
           "--use-build-mode",
+          "--working-dir",
+          process.cwd(),
           ...(await getTrapCachingExtractorConfigArgsForLang(config, language)),
           ...getExtractionVerbosityArguments(config.debugMode),
           ...getExtraOptionsFromEnv(["database", "trace-command"]),
@@ -731,7 +740,7 @@ export async function getCodeQLForCmd(
       const output = await runTool(cmd, codeqlArgs);
 
       try {
-        return JSON.parse(output);
+        return JSON.parse(output) as ResolveLanguagesOutput;
       } catch (e) {
         throw new Error(
           `Unexpected output from codeql resolve languages: ${e}`,
@@ -750,7 +759,7 @@ export async function getCodeQLForCmd(
       const output = await runTool(cmd, codeqlArgs);
 
       try {
-        return JSON.parse(output);
+        return JSON.parse(output) as BetterResolveLanguagesOutput;
       } catch (e) {
         throw new Error(
           `Unexpected output from codeql resolve languages with --format=betterjson: ${e}`,
@@ -774,7 +783,7 @@ export async function getCodeQLForCmd(
       const output = await runTool(cmd, codeqlArgs);
 
       try {
-        return JSON.parse(output);
+        return JSON.parse(output) as ResolveQueriesOutput;
       } catch (e) {
         throw new Error(`Unexpected output from codeql resolve queries: ${e}`);
       }
@@ -796,7 +805,7 @@ export async function getCodeQLForCmd(
       const output = await runTool(cmd, codeqlArgs);
 
       try {
-        return JSON.parse(output);
+        return JSON.parse(output) as ResolveBuildEnvironmentOutput;
       } catch (e) {
         throw new Error(
           `Unexpected output from codeql resolve build-environment: ${e} in\n${output}`,
@@ -1084,7 +1093,7 @@ export async function getCodeQLForCmd(
           },
         },
       ).exec();
-      return JSON.parse(extractorPath);
+      return JSON.parse(extractorPath) as string;
     },
     async mergeResults(
       sarifFiles: string[],
@@ -1338,7 +1347,7 @@ async function generateCodeScanningConfig(
 }
 
 function cloneObject<T>(obj: T): T {
-  return JSON.parse(JSON.stringify(obj));
+  return JSON.parse(JSON.stringify(obj)) as T;
 }
 
 // This constant sets the size of each TRAP cache in megabytes.