+++ src/config-utils.ts
@@ -976,7 +976,7 @@ export async function getConfig(
   const configString = fs.readFileSync(configFile, "utf8");
   logger.debug("Loaded config:");
   logger.debug(configString);
-  return JSON.parse(configString);
+  return JSON.parse(configString) as Config;
 }
 
 /**
@@ -1065,7 +1065,7 @@ function createRegistriesBlock(registries: RegistryConfigWithCredentials[]): {
  */
 export async function wrapEnvironment(
   env: Record<string, string | undefined>,
-  operation: Function,
+  operation: () => Promise<void>,
 ) {
   // Remember the original env
   const oldEnv = { ...process.env };
@@ -1106,6 +1106,16 @@ export async function parseBuildModeInput(
         BuildMode,
       ).join(", ")}.`,
     );
+  }
+
+  if (
+    languages.includes(Language.csharp) &&
+    (await features.getValue(Feature.DisableCsharpBuildless))
+  ) {
+    logger.warning(
+      "Scanning C# code without a build is temporarily unavailable. Falling back to 'autobuild' build mode.",
+    );
+    return BuildMode.Autobuild;
   }
 
   if (