+++ src/init-action.ts
@@ -26,6 +26,7 @@ import { EnvVar } from "./environment";
 import { Feature, Features } from "./feature-flags";
 import {
   checkInstallPython311,
+  cleanupDatabaseClusterDirectory,
   initCodeQL,
   initConfig,
   isSipEnabled,
@@ -321,6 +322,8 @@ async function run() {
   }
 
   try {
+    cleanupDatabaseClusterDirectory(config, logger);
+
     // Forward Go flags
     const goFlags = process.env["GOFLAGS"];
     if (goFlags) {
@@ -463,7 +466,7 @@ async function run() {
         );
       } else if (
         getTrapCachingEnabled() &&
-        (await features.getValue(Feature.CppTrapCachingEnabled, codeql))
+        (await codeQlVersionAtLeast(codeql, "2.17.5"))
       ) {
         logger.info("Enabling CodeQL C++ TRAP caching support");
         core.exportVariable(envVar, "true");